chore: move constants -> pkg
